bugbounty_7_agents_template/requirements.txt
aiohttp
pyyaml
rich
